// src/form-management/form.enum.ts
export enum Component {
  text = "text",
  textarea = "textarea",
  radioGroup = "radioGroup",
  select = "select",
  file = "file",
  group = "group"
}


export enum PropertyType {
  STRING = 'string',
  BOOLEAN = 'boolean',
  NUMBER = 'number',
  ARRAY= 'array',
  OBJECT = 'object',
  DATE = 'date',
}

export enum FormType {
  ADDRESS = 'address',
  EMAIL = 'email',
  PASSWORD = 'password',
  DATE = 'date',
}